refactor: 동시성 테스트 테스트명 등 수정
